Moved groupStates to separate file
