+++ widgets/sitesMap/src/SiteControl.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react';
 import type { DeploymentStatus, DeploymentStatusesSummary } from './types';
-import { DeploymentStatuses } from './types';
+import groupStates from './groupStates';
 
 interface SiteControlProps {
     site: {
@@ -73,29 +73,4 @@ const StateDescription: FunctionComponent<StateDescriptionProps> = ({ descriptio
             <strong>{value}</strong> {description}
         </span>
     );
-};
-
-type GroupState = { name: string; icon: string; colorSUI: string; severity: number; description: string };
-const groupStates: Record<DeploymentStatus, GroupState> = {
-    [DeploymentStatuses.Good]: {
-        name: 'good',
-        icon: 'checkmark',
-        colorSUI: 'green',
-        severity: 1,
-        description: ''
-    },
-    [DeploymentStatuses.InProgress]: {
-        name: 'in progress',
-        icon: 'spinner',
-        colorSUI: 'yellow',
-        severity: 2,
-        description: ''
-    },
-    [DeploymentStatuses.RequiresAttention]: {
-        name: 'failed',
-        icon: 'close',
-        colorSUI: 'red',
-        severity: 4,
-        description: ''
-    }
 };
